Fixed to not complain about deleting non-existant files Added TempFileFilter.createTempFolder()
git-svn-id: https://fluid.surelogic.com/svn/eng/trunk/common@15716 d02ae147-ee12-4a35-9452-91c89e8b2592

+++ src/com/surelogic/common/FileUtility.java
@@ -176,6 +176,9 @@ public final class FileUtility {
 				}
 			}
 		}
+		if (!path.exists()) {
+			return true; // Same result
+		}
 		success = path.delete();
 		if (!success) {
 			SLLogger.getLogger().warning(I18N.err(11, path.getAbsolutePath()));
@@ -212,6 +215,13 @@ public final class FileUtility {
 			return File.createTempFile(prefix, suffix);
 		}
 
+		public File createTempFolder() throws IOException {
+			File f = createTempFile();
+			f.delete();
+			f.mkdir();
+			return f;
+		}
+		
 		public boolean accept(File dir, String name) {
 			return name.startsWith(prefix) && name.endsWith(suffix);
 		}
